# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"elo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"username",
					"password",
					"elo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserInputDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"elo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"username",
					"password",
					"elo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserInputPatchDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"elo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"elo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"white": {
						"$ref": "#/components/schemas/UserOutputDTO"
					},
					"black": {
						"$ref": "#/components/schemas/UserOutputDTO"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"winner": {
						"$ref": "#/components/schemas/UserOutputDTO"
					},
					"hidden": {
						"type": "boolean"
					},
					"ranked": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"white",
					"black",
					"date",
					"winner",
					"hidden",
					"ranked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MoveOutputDTO": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/GameOutputDTO"
					},
					"move": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"move"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MoveInputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"move": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"move"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MoveInputPatchDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"move": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GameInputDTO": {
				"properties": {
					"white_id": {
						"type": "number",
						"format": "double"
					},
					"black_id": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"winner": {
						"type": "number",
						"format": "double"
					},
					"hidden": {
						"type": "boolean"
					},
					"ranked": {
						"type": "boolean"
					}
				},
				"required": [
					"white_id",
					"black_id",
					"date",
					"winner",
					"hidden",
					"ranked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameInputPatchDTO": {
				"properties": {
					"white": {
						"type": "number",
						"format": "double"
					},
					"black": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"winner_id": {
						"type": "number",
						"format": "double"
					},
					"hidden": {
						"type": "boolean"
					},
					"ranked": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticationInputDTO": {
				"properties": {
					"grant_type": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"grant_type",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "express-ts",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserInputDTO"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserInputPatchDTO"
							}
						}
					}
				}
			}
		},
		"/moves": {
			"get": {
				"operationId": "GetAllMoves",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MoveOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Move"
				],
				"security": [
					{
						"jwt": [
							"game:read"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MoveOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Move"
				],
				"security": [
					{
						"jwt": [
							"game:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MoveInputDTO"
							}
						}
					}
				}
			}
		},
		"/moves/{id}": {
			"get": {
				"operationId": "GetMoveById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MoveOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Move"
				],
				"security": [
					{
						"jwt": [
							"game:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMove",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Move"
				],
				"security": [
					{
						"jwt": [
							"game:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MoveOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Move"
				],
				"security": [
					{
						"jwt": [
							"game:create"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MoveInputPatchDTO"
							}
						}
					}
				}
			}
		},
		"/games": {
			"get": {
				"operationId": "GetAllGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:read"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "PostGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameInputDTO"
							}
						}
					}
				}
			}
		},
		"/games/{id}": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameInputPatchDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGame",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/user/{id}": {
			"get": {
				"operationId": "GetGameOfUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"jwt": [
							"game:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticationInputDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}